refactor(install): move error to install
