+++ src/extensions/install/failed-to-install.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Color } from "ink";
-import { PaperError } from "./paper-error";
+import { PaperError } from "../paper/exceptions";
 
 export class FailedToInstall extends PaperError {
   constructor(errorMessage: string){
