+++ src/extensions/paper/exceptions/index.ts
@@ -1,3 +1,2 @@
 export { AlreadyExistsError } from './already-exists';
-export { FailedToInstall } from './failed-to-install';
 export { PaperError } from './paper-error';
